text_compression/requirements.txt
py-lz4framed==0.14.0
msgpack==1.0.8
